ID-2976: Remove example container height
There was a static height set on a container, but overflow is not
handled, causing most content to grow above the set height.
This had the effect of the clipboard scrolling out of the view, because
it uses sticky positioning and the parent container element was
scrolling out of view.